microcontroller/requirements.txt
cffi==2.0.0
numpy==2.4.2
opencv-python==4.13.0.92
pycparser==3.0
scipy==1.17.1
sounddevice==0.5.5
vosk==0.3.45
pyttsx3==2.90
requests==2.32.3
python-dotenv==1.2.1